added state machine for loop animations.
We added a function to clear the color stops because that is what all
animations do in their own induvidual starting states.  We succesfully
started the loop function though have not programmed the delay. We
fixed all of the bugs.  We initialized the state machine at the very
end of start. We also succesfully implemented a delay state that would
correctly delay for the given number of milliseconds.