Fixed the remaining bugs with the hueIncrement.
The rainbow_pattern() function now behaves the way I want it to.  When
given (n, n), it lights up only one light; otherwise, it lights up the
range between the two numbers.  The previous incarnation of this
function was (silently!) dividing by 0 in some cases.

There's only one flaw left: this function does the right thing for about
20-30 seconds, then the light strings flash pink like mad and continue
flashing until reboot.  I'm not sure wat's causing it, but it's not
voltage, it's not amperage, and it's not due to some hidden need to
supply VCC and GND at the other end of the strip.