Eliminate the "crazy pink flash."
The cause was upate frequency.  Though our loop() deliberately waited
for a fixed amount of time before drawing each "rainbow frame", the loop
furiously idled for all the rest of the time.  That idling seems to
eventually cause the Arduino to send spurious data over the line.

The delay() function is generally avoided because it does nothing when
your code could be doing *something*, but in this case, doing nothing
for a few milliseconds every timeslice is the correct behavior.

I also moved some functions around and started adding comments.